[TF-17] Added buttons to the Create/Join Choice
The buttons only send information to the logger. I'm unsure of how to pass information between fragments (or switch fragments, for that matter) and I don't want to start doing it willy-nilly as that might leave sizable holes in the security implementation.
In any case, the comments should make it reasonably clear as to how this fragment is intended to work.